extras/python/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
defusedxml==0.7.1
idna==3.10
jira==3.8.0
joblib==1.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
Pygments==2.19.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
typing_extensions==4.12.2
urllib3==2.3.0
